ENG-1319 added implementation for uninstall, to regenerate the least of inconsistencies when uninstalling/rollbacking an updated bundle;

+++ src/main/java/org/entando/kubernetes/model/bundle/installable/CategoryInstallable.java
@@ -37,7 +37,9 @@ public class CategoryInstallable extends Installable<CategoryDescriptor> {
     public CompletableFuture<Void> uninstall() {
         return CompletableFuture.runAsync(() -> {
             log.info("Removing Label {}", getName());
-            engineService.deleteCategory(getName());
+            if (shouldCreate()) {
+                engineService.deleteCategory(getName());
+            }
         });
     }
